pyproject.toml
[project]
name = "mutliagent-orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Kiran",email = "Kiranpatil73827@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "openai (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
